Make autoboxing explicit
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpcore/trunk@1493732 13f79535-47bb-0310-9956-ffa450edef68

+++ httpcore-nio/src/main/java/org/apache/http/nio/pool/AbstractNIOConnPool.java
@@ -540,7 +540,7 @@ public abstract class AbstractNIOConnPool<T, C, E extends PoolEntry<T, C>>
     private int getMax(final T route) {
         final Integer v = this.maxPerRoute.get(route);
         if (v != null) {
-            return v;
+            return v.intValue();
         } else {
             return this.defaultMaxPerRoute;
         }
@@ -589,7 +589,7 @@ public abstract class AbstractNIOConnPool<T, C, E extends PoolEntry<T, C>>
         Args.positive(max, "Max value");
         this.lock.lock();
         try {
-            this.maxPerRoute.put(route, max);
+            this.maxPerRoute.put(route, Integer.valueOf(max));
         } finally {
             this.lock.unlock();
         }
