+++ httpcore/src/main/java/org/apache/http/pool/AbstractConnPool.java
@@ -319,7 +319,7 @@ public abstract class AbstractConnPool<T, C, E extends PoolEntry<T, C>>
     private int getMax(final T route) {
         final Integer v = this.maxPerRoute.get(route);
         if (v != null) {
-            return v;
+            return v.intValue();
         } else {
             return this.defaultMaxPerRoute;
         }
@@ -368,7 +368,7 @@ public abstract class AbstractConnPool<T, C, E extends PoolEntry<T, C>>
         Args.notNegative(max, "Max per route value");
         this.lock.lock();
         try {
-            this.maxPerRoute.put(route, max);
+            this.maxPerRoute.put(route, Integer.valueOf(max));
         } finally {
             this.lock.unlock();
         }
